Made sizing more flexible, added preview of two buttons
